requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
Calculator==1.1.0
colorama==0.4.4
execnet==1.7.1
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-assume==2.4.2
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.0
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
